// frontend/src/types/IUser.ts
export interface User {
    _id: string;
    id: string;
    name: string;
    email: string;
    profileImage?: string;
    role: "user" | "admin";
}

export interface RegisterData {
    name: string;
    email: string;
    password: string;
    confirmPassword: string;
}

export interface loginData {
    email: string;
    password: string;
}